Consolidate un-nested containers
