+++ amundsen_application/static/js/index.tsx
@@ -9,16 +9,14 @@ import { createStore, applyMiddleware } from 'redux';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import DocumentTitle from 'react-document-title';
 
-
-import AnnouncementPage from './containers/AnnouncementPage';
-import BrowsePage from "./containers/BrowsePage";
-import SearchPage from './containers/SearchPage';
-import TableDetail from './containers/TableDetail';
-
+import AnnouncementPage from './components/AnnouncementPage';
+import BrowsePage from './components/BrowsePage';
 import Feedback from './components/common/Feedback';
 import Footer from './components/Footer';
 import NavBar from './components/NavBar';
 import NotFoundPage from './components/NotFoundPage';
+import SearchPage from './components/SearchPage';
+import TableDetail from './components/TableDetail';
 
 import rootReducer from './ducks/rootReducer';
 import rootSaga from './ducks/rootSaga';
